# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Knowledge Base API",
    "version": "1.0.0",
    "description": "REST API สำหรับจัดการ Knowledge Base - ใช้ x-api-key header สำหรับ authentication"
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local" }
  ],
  "security": [{ "ApiKeyAuth": [] }],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/kb/knowledge/{id}": {
      "get": {
        "summary": "ดึง Knowledge Item ตาม ID",
        "operationId": "getKnowledge",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": { "200": { "description": "Knowledge item" } }
      },
      "put": {
        "summary": "แก้ไข Knowledge Item",
        "operationId": "updateKnowledge",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "question": { "type": "string", "maxLength": 500 },
                  "answer": { "type": "string" },
                  "type": { "type": "string", "enum": ["faq", "info", "procedure"] },
                  "categoryId": { "type": "string", "format": "uuid", "nullable": true },
                  "keywords": { "type": "array", "items": { "type": "string" } },
                  "tags": { "type": "array", "items": { "type": "string" } },
                  "isActive": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Updated knowledge item" } }
      },
      "delete": {
        "summary": "ลบ Knowledge Item",
        "operationId": "deleteKnowledge",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": { "200": { "description": "Deleted" } }
      }
    },
    "/api/v1/kb/knowledge/{id}/toggle": {
      "patch": {
        "summary": "เปิด/ปิดใช้งาน Knowledge Item",
        "operationId": "toggleKnowledge",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["isActive"],
                "properties": { "isActive": { "type": "boolean" } }
              }
            }
          }
        },
        "responses": { "200": { "description": "Toggled" } }
      }
    },
...
    "/api/v1/kb/flows": {
      "get": {
        "summary": "รายการ Flows",
        "operationId": "listFlows",
        "parameters": [
          { "name": "active_only", "in": "query", "schema": { "type": "boolean", "default": true } }
        ],
        "responses": { "200": { "description": "List of flows" } }
      },
      "post": {
        "summary": "สร้าง Flow ใหม่",
        "operationId": "createFlow",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string", "maxLength": 100 },
                  "description": { "type": "string", "nullable": true },
                  "triggerKeywords": { "type": "array", "items": { "type": "string" }, "default": [] },
                  "priority": { "type": "integer", "default": 0 },
                  "isActive": { "type": "boolean", "default": true }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created flow" } }
      }
    },
...
    "/api/v1/kb/flows/{id}/steps": {
      "post": {
        "summary": "เพิ่ม Step ใน Flow",
        "operationId": "createFlowStep",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["type", "message"],
                "properties": {
                  "type": { "type": "string", "enum": ["message", "options", "input", "end"] },
                  "message": { "type": "string" },
                  "options": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": ["label"],
                      "properties": {
                        "label": { "type": "string" },
                        "keywords": { "type": "array", "items": { "type": "string" } },
                        "nextStepId": { "type": "string", "format": "uuid", "nullable": true }
                      }
                    }
                  },
                  "nextStepId": { "type": "string", "format": "uuid", "nullable": true },
                  "isFinal": { "type": "boolean", "default": false }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created step" } }
      }
    },
    "/api/v1/kb/flows/{id}/steps/{stepId}": {
      "put": {
        "summary": "แก้ไข Flow Step",
        "operationId": "updateFlowStep",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "stepId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": { "type": "string", "enum": ["message", "options", "input", "end"] },
                  "message": { "type": "string" },
                  "options": { "type": "array" },
                  "nextStepId": { "type": "string", "format": "uuid", "nullable": true },
                  "isFinal": { "type": "boolean" },
                  "stepOrder": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Updated step" } }
      },
      "delete": {
        "summary": "ลบ Flow Step",
        "operationId": "deleteFlowStep",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "stepId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": { "200": { "description": "Deleted" } }
      }
    },
...
  }
}